chat-backend-done

+++ server/controllers/clientController.js
@@ -73,16 +73,45 @@ const getAllClients = async (req, res) => {
 
 }
 
+const getClientById = async (req,res)=>{
+    const {id} = req.params;
+    const user_id = req.user.id;
+
+    try {
+        const result = await pool.query(
+            'SELECT * FROM clients WHERE id = $1 AND user_id = $2',
+              [id, user_id]
+        );
+
+        if(result.rows.length === 0){
+            return res.status(404).json({
+                message : "NO CLIENT FOUND "
+            })
+        }
+
+        return res.status(200).json({
+            success: true,
+            client:result.rows[0]
+        })
+    } catch (error) {
+        return res.status(500).json({
+            message : error.message
+        });
+        
+    }
+}
+
 const updateClient = async (req, res) => {
     const { id } = req.params;
+    const user_id =  req.user.id;
     const { name, company_name, email, phone, address, city, country, website, industry, tax_id, payment_terms, hourly_rate, currency, status,
         client_type, notes } = req.body;
 
     try {
         const result = await pool.query(
-            'UPDATE clients SET name = $1 , company_name = $2 , email = $3 , phone=$4 ,address=$5 ,city=$6 ,country=$7, website=$8 , industry=$9 ,tax_id=$10 , payment_terms=$11 ,hourly_rate=$12 , currency=$13 , status =$14 ,client_type=$15, notes=$16 ,updated_at = NOW() WHERE id =$17  RETURNING * ',
+            'UPDATE clients SET name = $1 , company_name = $2 , email = $3 , phone=$4 ,address=$5 ,city=$6 ,country=$7, website=$8 , industry=$9 ,tax_id=$10 , payment_terms=$11 ,hourly_rate=$12 , currency=$13 , status =$14 ,client_type=$15, notes=$16 ,updated_at = NOW() WHERE id =$17 AND user_id = $18  RETURNING * ',
             [
-                name, company_name, email, phone, address, city, country, website, industry, tax_id, payment_terms, hourly_rate, currency, status, client_type, notes, id
+                name, company_name, email, phone, address, city, country, website, industry, tax_id, payment_terms, hourly_rate, currency, status, client_type, notes, id ,user_id
             ]
 
         )
@@ -105,10 +134,11 @@ const updateClient = async (req, res) => {
 
 const deleteClient = async (req, res) => {
     const { id } = req.params;
+    const user_id= req.user.id;
     try {
         const result = await pool.query(
-            'DELETE FROM clients WHERE id = $1  RETURNING *',
-            [id]
+            'DELETE FROM clients WHERE id = $1 AND user_id = $2 RETURNING *',
+            [id , user_id]
 
         )
         return res.status(200).json({
@@ -125,5 +155,5 @@ const deleteClient = async (req, res) => {
 }
 
 module.exports = {
-    createClient,getAllClients,updateClient,deleteClient
+    createClient,getAllClients,updateClient,deleteClient,getClientById
 };
